Do a refactor pass over the crypto module
