+++ config.cmake.in
@@ -66,6 +66,9 @@ foreach(component ${SOURCEMETA_CORE_COMPONENTS})
     if(@SOURCEMETA_CORE_CRYPTO_USE_SYSTEM_OPENSSL@)
       find_dependency(OpenSSL 3.0)
     endif()
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_preprocessor.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_numeric.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_text.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_crypto.cmake")
   elseif(component STREQUAL "regex")
     find_dependency(PCRE2 CONFIG)
@@ -111,6 +114,7 @@ foreach(component ${SOURCEMETA_CORE_COMPONENTS})
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_io.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_unicode.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_json.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_text.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_crypto.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_gzip.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_jsonl.cmake")
@@ -166,6 +170,9 @@ foreach(component ${SOURCEMETA_CORE_COMPONENTS})
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_semver.cmake")
   elseif(component STREQUAL "gzip")
     find_dependency(LibDeflate CONFIG)
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_preprocessor.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_numeric.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_text.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_crypto.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_core_gzip.cmake")
   elseif(component STREQUAL "html")
