+++ src/core/crypto/crypto_verify_rsa_other.cc
@@ -1,10 +1,8 @@
-#include <sourcemeta/core/crypto_sha256.h>
-#include <sourcemeta/core/crypto_sha384.h>
-#include <sourcemeta/core/crypto_sha512.h>
 #include <sourcemeta/core/crypto_verify.h>
+#include <sourcemeta/core/text.h>
 
 #include "crypto_bignum.h"
-#include "crypto_ecc.h"
+#include "crypto_verify_util.h"
 
 #include <array>       // std::array
 #include <cstddef>     // std::size_t
@@ -15,16 +13,6 @@
 #include <utility>     // std::unreachable
 
 namespace {
-
-constexpr std::size_t MAXIMUM_KEY_BYTES{512};
-
-auto strip_leading_zeros(std::string_view value) -> std::string_view {
-  while (!value.empty() && value.front() == '\x00') {
-    value.remove_prefix(1);
-  }
-
-  return value;
-}
 
 // The DigestInfo prefixes for the EMSA-PKCS1-v1_5 encoding, taken verbatim
 // from RFC 8017 Section 9.2 Note 1
@@ -55,33 +43,13 @@ auto digest_info_prefix(const sourcemeta::core::SignatureHashFunction hash)
   std::unreachable();
 }
 
-auto digest_message(const sourcemeta::core::SignatureHashFunction hash,
-                    const std::string_view message) -> std::string {
-  switch (hash) {
-    case sourcemeta::core::SignatureHashFunction::SHA256: {
-      const auto digest{sourcemeta::core::sha256_digest(message)};
-      return {reinterpret_cast<const char *>(digest.data()), digest.size()};
-    }
-    case sourcemeta::core::SignatureHashFunction::SHA384: {
-      const auto digest{sourcemeta::core::sha384_digest(message)};
-      return {reinterpret_cast<const char *>(digest.data()), digest.size()};
-    }
-    case sourcemeta::core::SignatureHashFunction::SHA512: {
-      const auto digest{sourcemeta::core::sha512_digest(message)};
-      return {reinterpret_cast<const char *>(digest.data()), digest.size()};
-    }
-  }
-
-  std::unreachable();
-}
-
 // EMSA-PKCS1-v1_5 encoding (RFC 8017 Section 9.2)
 auto build_encoded_message(const sourcemeta::core::SignatureHashFunction hash,
                            const std::string_view message,
                            const std::size_t key_length)
     -> std::optional<std::string> {
   const auto prefix{digest_info_prefix(hash)};
-  const auto digest{digest_message(hash, message)};
+  const auto digest{sourcemeta::core::digest_message(hash, message)};
   const auto encoded_length{prefix.size() + digest.size()};
 
   // RFC 8017 Section 9.2 step 3: "If emLen < tLen + 11, output 'intended
@@ -114,7 +82,7 @@ auto mask_generation(const sourcemeta::core::SignatureHashFunction hash,
     block.push_back(static_cast<char>((counter >> 16u) & 0xffu));
     block.push_back(static_cast<char>((counter >> 8u) & 0xffu));
     block.push_back(static_cast<char>(counter & 0xffu));
-    result.append(digest_message(hash, block));
+    result.append(sourcemeta::core::digest_message(hash, block));
     counter += 1;
   }
 
@@ -128,7 +96,7 @@ auto emsa_pss_verify(const sourcemeta::core::SignatureHashFunction hash,
                      const std::string_view message,
                      const std::string_view encoded_message,
                      const std::size_t encoded_bits) -> bool {
-  const auto digest{digest_message(hash, message)};
+  const auto digest{sourcemeta::core::digest_message(hash, message)};
   const auto hash_length{digest.size()};
   const auto salt_length{hash_length};
   const auto encoded_length{encoded_message.size()};
@@ -188,39 +156,9 @@ auto emsa_pss_verify(const sourcemeta::core::SignatureHashFunction hash,
   std::string verification_input(8, '\x00');
   verification_input.append(digest);
   verification_input.append(database.substr(database_length - salt_length));
-  const auto expected{digest_message(hash, verification_input)};
+  const auto expected{
+      sourcemeta::core::digest_message(hash, verification_input)};
   return expected == hash_value;
-}
-
-auto to_curve_parameters(const sourcemeta::core::EllipticCurve curve)
-    -> sourcemeta::core::EllipticCurveParameters {
-  switch (curve) {
-    case sourcemeta::core::EllipticCurve::P256:
-      return sourcemeta::core::curve_p256();
-    case sourcemeta::core::EllipticCurve::P384:
-      return sourcemeta::core::curve_p384();
-    case sourcemeta::core::EllipticCurve::P521:
-      return sourcemeta::core::curve_p521();
-  }
-
-  std::unreachable();
-}
-
-// FIPS 186-4 Section 6.4 step 2, deriving the integer e from the leftmost
-// bits of the message digest, truncated to the bit length of the order
-auto digest_to_integer(const sourcemeta::core::SignatureHashFunction hash,
-                       const std::string_view message,
-                       const std::size_t order_bits)
-    -> sourcemeta::core::Bignum {
-  const auto digest{digest_message(hash, message)};
-  auto value{sourcemeta::core::bignum_from_bytes(digest)};
-  const auto digest_bits{digest.size() * 8};
-  if (digest_bits > order_bits) {
-    value =
-        sourcemeta::core::bignum_shift_right(value, digest_bits - order_bits);
-  }
-
-  return value;
 }
 
 } // namespace
@@ -232,11 +170,11 @@ auto rsassa_pkcs1_v15_verify(const SignatureHashFunction hash,
                              const std::string_view exponent,
                              const std::string_view message,
                              const std::string_view signature) -> bool {
-  const auto stripped_modulus{strip_leading_zeros(modulus)};
-  const auto stripped_exponent{strip_leading_zeros(exponent)};
+  const auto stripped_modulus{sourcemeta::core::strip_left(modulus, '\x00')};
+  const auto stripped_exponent{sourcemeta::core::strip_left(exponent, '\x00')};
   if (stripped_modulus.empty() || stripped_exponent.empty() ||
-      stripped_modulus.size() > MAXIMUM_KEY_BYTES ||
-      stripped_exponent.size() > MAXIMUM_KEY_BYTES) {
+      stripped_modulus.size() > sourcemeta::core::MAXIMUM_KEY_BYTES ||
+      stripped_exponent.size() > sourcemeta::core::MAXIMUM_KEY_BYTES) {
     return false;
   }
 
@@ -270,11 +208,11 @@ auto rsassa_pss_verify(const SignatureHashFunction hash,
                        const std::string_view exponent,
                        const std::string_view message,
                        const std::string_view signature) -> bool {
-  const auto stripped_modulus{strip_leading_zeros(modulus)};
-  const auto stripped_exponent{strip_leading_zeros(exponent)};
+  const auto stripped_modulus{sourcemeta::core::strip_left(modulus, '\x00')};
+  const auto stripped_exponent{sourcemeta::core::strip_left(exponent, '\x00')};
   if (stripped_modulus.empty() || stripped_exponent.empty() ||
-      stripped_modulus.size() > MAXIMUM_KEY_BYTES ||
-      stripped_exponent.size() > MAXIMUM_KEY_BYTES) {
+      stripped_modulus.size() > sourcemeta::core::MAXIMUM_KEY_BYTES ||
+      stripped_exponent.size() > sourcemeta::core::MAXIMUM_KEY_BYTES) {
     return false;
   }
 
@@ -315,75 +253,6 @@ auto rsassa_pss_verify(const SignatureHashFunction hash,
   const auto encoded_message{std::string_view{full_representative}.substr(
       key_length - encoded_length)};
   return emsa_pss_verify(hash, message, encoded_message, encoded_bits);
-}
-
-auto ecdsa_verify(const EllipticCurve curve, const SignatureHashFunction hash,
-                  const std::string_view coordinate_x,
-                  const std::string_view coordinate_y,
-                  const std::string_view message,
-                  const std::string_view signature) -> bool {
-  const auto parameters{to_curve_parameters(curve)};
-  const auto field_bytes{parameters.field_bytes};
-
-  // RFC 7518 Section 3.4: the signature is the fixed-width concatenation of
-  // the two integers, each as long as the curve field
-  if (signature.size() != field_bytes * 2) {
-    return false;
-  }
-
-  const auto r{bignum_from_bytes(signature.substr(0, field_bytes))};
-  const auto s{bignum_from_bytes(signature.substr(field_bytes))};
-
-  // FIPS 186-4 Section 6.4.2 step 1: both integers must lie in [1, n - 1]
-  if (bignum_is_zero(r) || bignum_compare(r, parameters.order) >= 0 ||
-      bignum_is_zero(s) || bignum_compare(s, parameters.order) >= 0) {
-    return false;
-  }
-
-  // Reject coordinates wider than the field, matching the platform backends
-  // and preventing an oversized input from being truncated into a valid key
-  const auto stripped_x{strip_leading_zeros(coordinate_x)};
-  const auto stripped_y{strip_leading_zeros(coordinate_y)};
-  if (stripped_x.size() > field_bytes || stripped_y.size() > field_bytes) {
-    return false;
-  }
-
-  const auto public_x{bignum_from_bytes(stripped_x)};
-  const auto public_y{bignum_from_bytes(stripped_y)};
-
-  // The public key must be a valid point: coordinates below the field prime
-  // and satisfying the curve equation
-  if (bignum_compare(public_x, parameters.prime) >= 0 ||
-      bignum_compare(public_y, parameters.prime) >= 0 ||
-      !point_on_curve(public_x, public_y, parameters)) {
-    return false;
-  }
-
-  const auto order_bits{bignum_bit_length(parameters.order)};
-  const auto digest_integer{digest_to_integer(hash, message, order_bits)};
-  const auto s_inverse{bignum_mod_inverse(s, parameters.order)};
-  const auto u1{
-      bignum_mod_multiply(digest_integer, s_inverse, parameters.order)};
-  const auto u2{bignum_mod_multiply(r, s_inverse, parameters.order)};
-
-  const JacobianPoint generator{parameters.generator_x, parameters.generator_y,
-                                bignum_from_u64(1)};
-  const JacobianPoint public_point{public_x, public_y, bignum_from_u64(1)};
-  const auto point{point_add(
-      point_scalar_multiply(u1, generator, parameters),
-      point_scalar_multiply(u2, public_point, parameters), parameters)};
-
-  // FIPS 186-4 Section 6.4.2 step 6: reject when the combination is the
-  // point at infinity
-  if (point_is_infinity(point)) {
-    return false;
-  }
-
-  // FIPS 186-4 Section 6.4.2 step 7: the signature is valid when the affine
-  // x coordinate, reduced modulo the order, equals r
-  auto candidate{point_affine_x(point, parameters)};
-  bignum_reduce(candidate, parameters.order);
-  return bignum_compare(candidate, r) == 0;
 }
 
 } // namespace sourcemeta::core
