+++ test/crypto/pyca_cryptography_test.cc
@@ -1,0 +1,411 @@
+#include <gtest/gtest.h>
+
+#include <sourcemeta/core/crypto.h>
+#include <sourcemeta/core/io.h>
+#include <sourcemeta/core/text.h>
+
+#include <cstddef>     // std::size_t
+#include <cstdint>     // std::uint64_t
+#include <filesystem>  // std::filesystem::path
+#include <functional>  // std::function
+#include <optional>    // std::optional, std::nullopt
+#include <string>      // std::string
+#include <string_view> // std::string_view
+#include <utility>     // std::move
+
+// Every NIST CAVP response file ignores blank lines and "#" comments, leaving
+// section headers and "Key = Value" records
+template <typename Callback>
+static auto for_each_vector_line(const std::filesystem::path &path,
+                                 Callback callback) -> void {
+  auto stream{sourcemeta::core::read_file(path)};
+  sourcemeta::core::for_each_line(stream,
+                                  [&callback](const std::string_view line) {
+                                    if (line.empty() || line.front() == '#') {
+                                      return;
+                                    }
+
+                                    callback(line);
+                                  });
+}
+
+static auto to_signature_hash_function(const std::string_view name)
+    -> std::optional<sourcemeta::core::SignatureHashFunction> {
+  if (name == "SHA256" || name == "SHA-256") {
+    return sourcemeta::core::SignatureHashFunction::SHA256;
+  } else if (name == "SHA384" || name == "SHA-384") {
+    return sourcemeta::core::SignatureHashFunction::SHA384;
+  } else if (name == "SHA512" || name == "SHA-512") {
+    return sourcemeta::core::SignatureHashFunction::SHA512;
+  } else {
+    return std::nullopt;
+  }
+}
+
+struct CurveParameters {
+  sourcemeta::core::EllipticCurve curve;
+  std::size_t coordinate_bytes;
+};
+
+static auto to_curve(const std::string_view name)
+    -> std::optional<CurveParameters> {
+  if (name == "P-256") {
+    return CurveParameters{sourcemeta::core::EllipticCurve::P256, 32};
+  } else if (name == "P-384") {
+    return CurveParameters{sourcemeta::core::EllipticCurve::P384, 48};
+  } else if (name == "P-521") {
+    return CurveParameters{sourcemeta::core::EllipticCurve::P521, 66};
+  } else {
+    return std::nullopt;
+  }
+}
+
+// The hash functions are overloaded, so the registrars name the digest variant
+// that returns a hex string through this alias
+using DigestFunction = auto (*)(const std::string_view) -> std::string;
+
+// Every vector in this file reduces to running a closure that asserts on data
+// captured at registration time, so a single fixture carrying that closure
+// stands in for every mechanism
+class PyCA_Cryptography_Test : public testing::Test {
+public:
+  explicit PyCA_Cryptography_Test(std::function<void()> input_assertion)
+      : assertion{std::move(input_assertion)} {}
+
+  auto TestBody() -> void override { this->assertion(); }
+
+private:
+  const std::function<void()> assertion;
+};
+
+static auto register_case(const std::string &suite_name,
+                          const std::string &test_name,
+                          std::function<void()> assertion) -> void {
+  testing::RegisterTest(
+      suite_name.c_str(), test_name.c_str(), nullptr, nullptr, __FILE__,
+      __LINE__, [assertion = std::move(assertion)]() -> testing::Test * {
+        return new PyCA_Cryptography_Test(assertion);
+      });
+}
+
+static auto register_sha_msg_tests(const std::filesystem::path &file_path,
+                                   const std::string &suite_name,
+                                   const DigestFunction digest) -> void {
+  std::string current_length;
+  std::string current_message_hex;
+
+  for_each_vector_line(file_path, [&](const std::string_view line) {
+    if (line.front() == '[') {
+      return;
+    }
+
+    const auto field{
+        sourcemeta::core::split_once(line, std::string_view{" = "})};
+    if (!field.has_value()) {
+      return;
+    }
+
+    if (field->first == "Len") {
+      current_length = field->second;
+    } else if (field->first == "Msg") {
+      current_message_hex = field->second;
+    } else if (field->first == "MD") {
+      const auto expected_digest{std::string{field->second}};
+      const auto empty_message{current_length == "0"};
+      const auto message_hex{current_message_hex};
+
+      register_case(suite_name, "len_" + current_length, [=]() {
+        const auto message{
+            empty_message
+                ? std::string{}
+                : sourcemeta::core::hex_to_bytes(message_hex).value()};
+        EXPECT_EQ(digest(message), expected_digest);
+      });
+    }
+  });
+}
+
+static auto register_sha_monte_tests(const std::filesystem::path &file_path,
+                                     const std::string &suite_name,
+                                     const DigestFunction digest) -> void {
+  std::string seed;
+  std::string current_count;
+
+  for_each_vector_line(file_path, [&](const std::string_view line) {
+    if (line.front() == '[') {
+      return;
+    }
+
+    const auto field{
+        sourcemeta::core::split_once(line, std::string_view{" = "})};
+    if (!field.has_value()) {
+      return;
+    }
+
+    if (field->first == "Seed") {
+      seed = field->second;
+    } else if (field->first == "COUNT") {
+      current_count = field->second;
+    } else if (field->first == "MD") {
+      const auto expected_digest{std::string{field->second}};
+      const auto current_seed{seed};
+
+      register_case(suite_name, "count_" + current_count, [=]() {
+        // NIST SHAVS Monte Carlo pseudocode:
+        // MD[0] = MD[1] = MD[2] = Seed
+        // For i = 3 to 1002:
+        //   M[i] = MD[i-3] || MD[i-2] || MD[i-1]
+        //   MD[i] = SHA(M[i])
+        // Output MD[1002]
+        auto md_0{current_seed};
+        auto md_1{current_seed};
+        auto md_2{current_seed};
+        for (std::uint64_t iteration{3}; iteration <= 1002; ++iteration) {
+          const auto message{sourcemeta::core::hex_to_bytes(md_0).value() +
+                             sourcemeta::core::hex_to_bytes(md_1).value() +
+                             sourcemeta::core::hex_to_bytes(md_2).value()};
+          auto next{digest(message)};
+          md_0 = std::move(md_1);
+          md_1 = std::move(md_2);
+          md_2 = std::move(next);
+        }
+
+        EXPECT_EQ(md_2, expected_digest);
+      });
+
+      // The output becomes the seed for the next checkpoint
+      seed = expected_digest;
+    }
+  });
+}
+
+static auto register_sha_tests(const std::filesystem::path &directory,
+                               const std::string &name,
+                               const DigestFunction digest) -> void {
+  register_sha_msg_tests(directory / (name + "ShortMsg.rsp"),
+                         "PyCA_Cryptography_" + name + "_ShortMsg", digest);
+  register_sha_msg_tests(directory / (name + "LongMsg.rsp"),
+                         "PyCA_Cryptography_" + name + "_LongMsg", digest);
+  register_sha_monte_tests(directory / (name + "Monte.rsp"),
+                           "PyCA_Cryptography_" + name + "_Monte", digest);
+}
+
+struct RSAVector {
+  std::string modulus_bits;
+  std::string modulus_hex;
+  std::string sha_algorithm;
+  std::string exponent_hex;
+  std::string message_hex;
+  std::string signature_hex;
+};
+
+template <typename Callback>
+static auto for_each_rsa_case(const std::filesystem::path &file_path,
+                              Callback callback) -> void {
+  RSAVector vector;
+  std::uint64_t case_count{0};
+
+  for_each_vector_line(file_path, [&](const std::string_view line) {
+    if (line.front() == '[') {
+      // The section header carries the modulus bit size, like "[mod = 1024]"
+      const auto header{sourcemeta::core::strip_right(
+          sourcemeta::core::strip_left(line, '['), ']')};
+      const auto field{
+          sourcemeta::core::split_once(header, std::string_view{" = "})};
+      if (field.has_value()) {
+        vector.modulus_bits = field->second;
+      }
+
+      return;
+    }
+
+    const auto field{
+        sourcemeta::core::split_once(line, std::string_view{" = "})};
+    if (!field.has_value()) {
+      return;
+    }
+
+    if (field->first == "n") {
+      vector.modulus_hex = field->second;
+    } else if (field->first == "SHAAlg") {
+      vector.sha_algorithm = field->second;
+    } else if (field->first == "e") {
+      vector.exponent_hex = field->second;
+    } else if (field->first == "Msg") {
+      vector.message_hex = field->second;
+    } else if (field->first == "S") {
+      vector.signature_hex = field->second;
+    } else if (field->first == "Result") {
+      case_count += 1;
+      callback(vector, field->second, case_count);
+    }
+  });
+}
+
+static auto register_rsa_sigver15_tests(const std::filesystem::path &file_path)
+    -> void {
+  for_each_rsa_case(file_path, [](const RSAVector &vector,
+                                  const std::string_view result,
+                                  const std::uint64_t case_count) {
+    const auto hash{to_signature_hash_function(vector.sha_algorithm)};
+    if (!hash.has_value()) {
+      return;
+    }
+
+    const auto hash_function{hash.value()};
+    const auto expected{result.starts_with("P")};
+    const auto test_name{
+        "mod" + vector.modulus_bits + "_" + vector.sha_algorithm + "_case_" +
+        std::to_string(case_count) + "_" + (expected ? "pass" : "fail")};
+    register_case(
+        "PyCA_Cryptography_RSA_SigVer15", test_name,
+        [vector, hash_function, expected]() {
+          EXPECT_EQ(
+              sourcemeta::core::rsassa_pkcs1_v15_verify(
+                  hash_function,
+                  sourcemeta::core::hex_to_bytes(vector.modulus_hex).value(),
+                  sourcemeta::core::hex_to_bytes(vector.exponent_hex).value(),
+                  sourcemeta::core::hex_to_bytes(vector.message_hex).value(),
+                  sourcemeta::core::hex_to_bytes(vector.signature_hex).value()),
+              expected);
+        });
+  });
+}
+
+static auto register_rsa_sigverpss_tests(const std::filesystem::path &file_path)
+    -> void {
+  // Every case in this suite uses a 10-byte salt, while RFC 7518 Section 3.5
+  // fixes the salt size to the hash function output ("The size of the salt
+  // value is the same size as the hash function output"). Verification must
+  // therefore reject every case, including the ones whose signatures are
+  // otherwise cryptographically valid. The suite still exercises the full
+  // unmasking and consistency checking paths with realistic data
+  for_each_rsa_case(file_path, [](const RSAVector &vector,
+                                  const std::string_view,
+                                  const std::uint64_t case_count) {
+    const auto hash{to_signature_hash_function(vector.sha_algorithm)};
+    if (!hash.has_value()) {
+      return;
+    }
+
+    const auto hash_function{hash.value()};
+    const auto test_name{"mod" + vector.modulus_bits + "_" +
+                         vector.sha_algorithm + "_case_" +
+                         std::to_string(case_count)};
+    register_case(
+        "PyCA_Cryptography_RSA_SigVerPSS", test_name,
+        [vector, hash_function]() {
+          EXPECT_FALSE(sourcemeta::core::rsassa_pss_verify(
+              hash_function,
+              sourcemeta::core::hex_to_bytes(vector.modulus_hex).value(),
+              sourcemeta::core::hex_to_bytes(vector.exponent_hex).value(),
+              sourcemeta::core::hex_to_bytes(vector.message_hex).value(),
+              sourcemeta::core::hex_to_bytes(vector.signature_hex).value()));
+        });
+  });
+}
+
+static auto register_ecdsa_sigver_tests(const std::filesystem::path &file_path)
+    -> void {
+  std::optional<CurveParameters> current_curve;
+  std::optional<sourcemeta::core::SignatureHashFunction> current_hash;
+  std::string message_hex;
+  std::string coordinate_x_hex;
+  std::string coordinate_y_hex;
+  std::string r_hex;
+  std::string s_hex;
+  std::uint64_t case_count{0};
+
+  for_each_vector_line(file_path, [&](const std::string_view line) {
+    if (line.front() == '[') {
+      // The section header pairs a curve and a hash, like "[P-256,SHA-256]"
+      const auto header{sourcemeta::core::strip_right(
+          sourcemeta::core::strip_left(line, '['), ']')};
+      const auto parts{sourcemeta::core::split_once(header, ',')};
+      if (parts.has_value()) {
+        current_curve = to_curve(parts->first);
+        current_hash = to_signature_hash_function(parts->second);
+      }
+
+      return;
+    }
+
+    const auto field{
+        sourcemeta::core::split_once(line, std::string_view{" = "})};
+    if (!field.has_value()) {
+      return;
+    }
+
+    if (field->first == "Msg") {
+      message_hex = field->second;
+    } else if (field->first == "Qx") {
+      coordinate_x_hex = field->second;
+    } else if (field->first == "Qy") {
+      coordinate_y_hex = field->second;
+    } else if (field->first == "R") {
+      r_hex = field->second;
+    } else if (field->first == "S") {
+      s_hex = field->second;
+    } else if (field->first == "Result") {
+      case_count += 1;
+      if (!current_curve.has_value() || !current_hash.has_value()) {
+        return;
+      }
+
+      const auto width{current_curve.value().coordinate_bytes};
+      const auto curve{current_curve.value().curve};
+      const auto hash{current_hash.value()};
+      const auto expected{field->second.starts_with("P")};
+      const auto test_name{"case_" + std::to_string(case_count) + "_" +
+                           (expected ? "pass" : "fail")};
+      register_case("PyCA_Cryptography_ECDSA_SigVer", test_name, [=]() {
+        // The vectors elide leading zero nibbles, so an odd hex length means a
+        // single zero was dropped, and each coordinate is padded to the curve
+        // width
+        const auto coordinate_x{sourcemeta::core::pad_left(
+            sourcemeta::core::hex_to_bytes(coordinate_x_hex, true).value(),
+            width, '\x00')};
+        const auto coordinate_y{sourcemeta::core::pad_left(
+            sourcemeta::core::hex_to_bytes(coordinate_y_hex, true).value(),
+            width, '\x00')};
+        const auto message{
+            sourcemeta::core::hex_to_bytes(message_hex, true).value()};
+        const auto signature{
+            sourcemeta::core::pad_left(
+                sourcemeta::core::hex_to_bytes(r_hex, true).value(), width,
+                '\x00') +
+            sourcemeta::core::pad_left(
+                sourcemeta::core::hex_to_bytes(s_hex, true).value(), width,
+                '\x00')};
+        EXPECT_EQ(sourcemeta::core::ecdsa_verify(curve, hash, coordinate_x,
+                                                 coordinate_y, message,
+                                                 signature),
+                  expected);
+      });
+    }
+  });
+}
+
+auto main(int argc, char **argv) -> int {
+  testing::InitGoogleTest(&argc, argv);
+  const std::filesystem::path suite_path{PYCA_CRYPTOGRAPHY_PATH};
+  const auto hashes_path{suite_path / "hashes"};
+
+  register_sha_tests(hashes_path / "SHA1", "SHA1",
+                     static_cast<DigestFunction>(sourcemeta::core::sha1));
+  register_sha_tests(hashes_path / "SHA2", "SHA256",
+                     static_cast<DigestFunction>(sourcemeta::core::sha256));
+  register_sha_tests(hashes_path / "SHA2", "SHA384",
+                     static_cast<DigestFunction>(sourcemeta::core::sha384));
+  register_sha_tests(hashes_path / "SHA2", "SHA512",
+                     static_cast<DigestFunction>(sourcemeta::core::sha512));
+
+  const auto rsa_path{suite_path / "asymmetric" / "RSA" / "FIPS_186-2"};
+  register_rsa_sigver15_tests(rsa_path / "SigVer15_186-3.rsp");
+  register_rsa_sigverpss_tests(rsa_path / "SigVerPSS_186-3.rsp");
+
+  register_ecdsa_sigver_tests(suite_path / "asymmetric" / "ECDSA" /
+                              "FIPS_186-3" / "SigVer.rsp");
+
+  return RUN_ALL_TESTS();
+}
