+++ test/text/CMakeLists.txt
@@ -4,7 +4,9 @@ sourcemeta_googletest(NAMESPACE sourcemeta PROJECT core NAME text
           text_truncate_test.cc text_remove_suffix_ignore_case_test.cc
           text_trim_test.cc text_take_until_test.cc text_split_once_test.cc
           text_split_test.cc text_join_to_test.cc
-          text_hex_to_bytes_test.cc)
+          text_hex_to_bytes_test.cc text_bytes_to_hex_test.cc
+          text_strip_left_test.cc text_strip_right_test.cc
+          text_pad_left_test.cc)
 
 target_link_libraries(sourcemeta_core_text_unit
   PRIVATE sourcemeta::core::text)
