+++ test/text/text_pad_left_test.cc
@@ -1,0 +1,34 @@
+#include <gtest/gtest.h>
+
+#include <sourcemeta/core/text.h>
+
+#include <string> // std::string
+
+TEST(Text_pad_left, pads_to_width) {
+  EXPECT_EQ(sourcemeta::core::pad_left("42", 5, '0'), "00042");
+}
+
+TEST(Text_pad_left, already_at_width) {
+  EXPECT_EQ(sourcemeta::core::pad_left("hello", 5, '0'), "hello");
+}
+
+TEST(Text_pad_left, longer_than_width) {
+  EXPECT_EQ(sourcemeta::core::pad_left("hello", 3, '0'), "hello");
+}
+
+TEST(Text_pad_left, single_character_pad) {
+  EXPECT_EQ(sourcemeta::core::pad_left("x", 4, ' '), "   x");
+}
+
+TEST(Text_pad_left, empty_input) {
+  EXPECT_EQ(sourcemeta::core::pad_left("", 3, '0'), "000");
+}
+
+TEST(Text_pad_left, zero_width) {
+  EXPECT_EQ(sourcemeta::core::pad_left("abc", 0, '0'), "abc");
+}
+
+TEST(Text_pad_left, pads_with_nul_bytes) {
+  const std::string expected("\x00\x00\x2a", 3);
+  EXPECT_EQ(sourcemeta::core::pad_left("\x2a", 3, '\x00'), expected);
+}
